% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{binary_tree_1}
\alias{binary_tree_1}
\title{An example dataset with a tree-shaped trajectory}
\format{
raw_data, raw_normalized and clustering
}
\source{
\url{https://doi.org/10.5281/zenodo.1443566}
}
\usage{
data(binary_tree_1)
}
\description{
An example dataset with a tree-shaped trajectory
}
\examples{
data(binary_tree_1)
}
\keyword{datasets}
