% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrMass-data.R
\docType{data}
\name{bmrMass}
\alias{bmrMass}
\title{Basic metabolic rate/body mass data}
\format{
A data.frame with 549 observations of 2 variables (BMR, and Mass)
}
\usage{
data(bmrMass)
}
\description{
Data from Genoud et al. (2018) for the basic metabolic rate and body mass of
549 mammals.
}
\examples{
data(bmrMass)
## testing Kleiber's law
i.test(BMR ~ I(Mass^(3/4)), data = bmrMass)
}
\keyword{dataset}
