% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ChangeTrajRoot}
\alias{ChangeTrajRoot}
\alias{ChangeTrajRoot.SingleCellExperiment}
\alias{ChangeTrajRoot,SingleCellExperiment-method}
\title{Change the root cluster of a trajectory}
\usage{
ChangeTrajRoot.SingleCellExperiment(object, traj.name, root.cluster)

\S4method{ChangeTrajRoot}{SingleCellExperiment}(object, traj.name, root.cluster)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{traj.name}{a character for the name of the trajectory}

\item{root.cluster}{an integer for the name of the new root cluster}
}
\value{
an object of class \code{SingleCellExperiment}
}
\description{
The function changes the root cluster of a trajectory, which is initially
chosen randomly.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce,selection.method = 1,selection.N.models = 2)
sce <- RunSmoothing(sce)
clustering_name <-  ReturnTrajNames(sce)[1]
sce <- ChangeTrajRoot(sce,traj.name=clustering_name,root.cluster=2)
}
\keyword{cluster}
\keyword{node}
\keyword{root}
\keyword{start}
\keyword{trajectory}
