% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ClusterCentroids}
\alias{ClusterCentroids}
\title{Calculate cluster centroids}
\usage{
ClusterCentroids(clustering, X)
}
\arguments{
\item{clustering}{a vector of cluster labels.}

\item{X}{a matrix that specifies the embedding (dim.reduced data).}
}
\value{
a matrix
}
\description{
The function calculates the cluster centroids.
}
\keyword{center}
\keyword{centroids}
\keyword{cluster}
