% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Connectivity}
\alias{Connectivity}
\title{Calculate connectivity}
\usage{
Connectivity(clusterings, mst.list)
}
\arguments{
\item{clusterings}{a named list of clustering results.}

\item{mst.list}{a named list of MSTs}
}
\value{
a list of MSTs
}
\description{
The function calculates the node-specific connectivity 
for each MST in a list.
}
\keyword{MST}
\keyword{ape}
