% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{CreateMilestoneNetwork}
\alias{CreateMilestoneNetwork}
\title{Create milestone network from lineages}
\usage{
CreateMilestoneNetwork(lineages)
}
\arguments{
\item{lineages}{a list of vectors that specify the lineages.}
}
\value{
a data frame
}
\description{
The function creates a milestone network from lineages.
}
\keyword{lineage}
\keyword{milestone}
\keyword{network}
