% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{CreateProgressions}
\alias{CreateProgressions}
\title{Create milestone progressions}
\usage{
CreateProgressions(clustering, X, milestone.network)
}
\arguments{
\item{clustering}{a vector of cluster labels.}

\item{X}{a matrix that specifies the embedding (dim.reduced data).}

\item{milestone.network}{a data frame of the milestone network.}
}
\value{
a data frame
}
\description{
The function creates the progressions for the dynwrap object.
}
\keyword{milestone}
\keyword{progressions}
