% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{FindLineages}
\alias{FindLineages}
\title{Find lineages of an MST from a starting cluster}
\usage{
FindLineages(mst.out, start.cluster)
}
\arguments{
\item{mst.out}{an object of class "mst"}

\item{start.cluster}{a character or numeric denoting the starting cluster.}
}
\value{
a list that includes the lineages
}
\description{
The function finds all lineages of an MST. The starting cluster needs
to be provided.
}
\keyword{MST}
\keyword{lineage}
