% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProjector.R
\name{iProjector}
\alias{iProjector}
\title{iProjector}
\usage{
iProjector(
  C,
  targets,
  v = NULL,
  tolerance = .Machine$double.eps,
  maxIter = 10000L
)
}
\arguments{
\item{C}{a matrix with D constraints in the rows for |E| entities in the columns, must be of full row rank}

\item{targets}{a matrix/vector of with D rows/entries with the expected values for the constraints in C}

\item{v}{the reference distribution, defaults to the uniform distribution}

\item{maxIter}{the maximal number of iterations, defaults to 10 000}

\item{eps}{the tolerance}
}
\value{
a list with following entries
\tabular{ll}{
 \code{p} \tab the I-projection of the reference distribution v onto the
   TOTEMplex defined by C \cr
 \code{converged} \tab a logical indicating convergence (\code{TRUE}) or
   failure (\code{FALSE}) \cr
 \code{info} \tab information about the convergence
 }
}
\description{
Compute the I-projection given the coefficient matrix \code{C} and expected values in \code{targets}
}
