% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{MergeClusters}
\alias{MergeClusters}
\alias{MergeClusters.SingleCellExperiment}
\alias{MergeClusters,SingleCellExperiment-method}
\title{Merge clusters of a clustering}
\usage{
MergeClusters.SingleCellExperiment(
  object,
  clustering.name,
  merged.clusters,
  merged.name
)

\S4method{MergeClusters}{SingleCellExperiment}(object, clustering.name, merged.clusters, merged.name)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{clustering.name}{a character for the name of the clustering.}

\item{merged.clusters}{a vector denoting the names of the clusters to merge.}

\item{merged.name}{a character denoting the name of the merged clusters.
By default, the new name will be the name of the cluster that appears first
in the clustering.}
}
\value{
an object of class \code{SingleCellExperiment}
}
\description{
The function merges clusters of a clustering. The original clustering will be
replaced by the new one, and the only way to get it back is to re-run the 
\code{RunClustering} function.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) #Use default N.clusterings=10000
sce <- SelectClusterings(sce,selection.method = 1,selection.N.models = 2)
clustering_name <-  ReturnTrajNames(sce)[1]
sce <- MergeClusters(sce,clustering.name=clustering_name,
                     merged.clusters=c(1,2),merged.name=c(1))
}
\keyword{clustering}
\keyword{clusters}
\keyword{merge}
