% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{PrepareTotem}
\alias{PrepareTotem}
\alias{PrepareTotem.SingleCellExperiment}
\alias{PrepareTotem,SingleCellExperiment-method}
\title{Prepare \code{SingleCellExperiment} object for \code{Totem} analysis}
\usage{
PrepareTotem.SingleCellExperiment(object)

\S4method{PrepareTotem}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}
}
\value{
an object of class \code{SingleCellExperiment}
}
\description{
The function prepares the \code{SingleCellExperiment} object for
\code{Totem} analysis. The only required input is an object of class
\code{SingleCellExperiment} with data stored at least in the \code{logcounts}
 and \code{counts} slots.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)

}
\keyword{data}
\keyword{normalized}
\keyword{prepare}
\keyword{raw}
\keyword{totem}
