% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ReturnClustering}
\alias{ReturnClustering}
\alias{ReturnClustering.SingleCellExperiment}
\alias{ReturnClustering,SingleCellExperiment-method}
\title{Return a clustering}
\usage{
ReturnClustering.SingleCellExperiment(object, clustering.name)

\S4method{ReturnClustering}{SingleCellExperiment}(object, clustering.name)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{clustering.name}{a character for the clustering name.}
}
\value{
a named character vector
}
\description{
The function returns a clustering corresponding 
to the provided clustering name.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce,selection.method = 1,selection.N.models = 2)
clustering_name <-  ReturnTrajNames(sce)[1]
clustering <- ReturnClustering(sce,clustering_name)
}
\keyword{clustering}
