% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ReturnDynwrapObject}
\alias{ReturnDynwrapObject}
\alias{ReturnDynwrapObject.SingleCellExperiment}
\alias{ReturnDynwrapObject,SingleCellExperiment-method}
\title{Return a trajectory as a dynwrap object}
\usage{
ReturnDynwrapObject.SingleCellExperiment(object, traj.name)

\S4method{ReturnDynwrapObject}{SingleCellExperiment}(object, traj.name)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{traj.name}{a character for the trajectory name}
}
\value{
an object of class \code{dynwrap}
}
\description{
The function returns a Slingshot-smoothed trajectory 
as a \code{dynwrap} object
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce,selection.method = 1,selection.N.models = 2)
sce <- RunSmoothing(sce)
clustering_name <-  ReturnTrajNames(sce)[1]
dynwrap_object <- ReturnDynwrapObject(sce,clustering_name)
}
\keyword{dynwrap}
\keyword{smooth}
\keyword{trajectory}
