% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ReturnMSTNetwork}
\alias{ReturnMSTNetwork}
\alias{ReturnMSTNetwork.SingleCellExperiment}
\alias{ReturnMSTNetwork,SingleCellExperiment-method}
\title{Return an MST network of a clustering}
\usage{
ReturnMSTNetwork.SingleCellExperiment(object, clustering.name)

\S4method{ReturnMSTNetwork}{SingleCellExperiment}(object, clustering.name)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{clustering.name}{a character for the clustering name.}
}
\value{
an object of class "mst"
}
\description{
The function returns the MST network of a clustering.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
clustering_name <-  "3.1"
mst_clustering <- ReturnMSTNetwork(sce,clustering_name)
}
\keyword{MST}
\keyword{clustering}
\keyword{network}
