% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ReturnSmoothedTrajNetwork}
\alias{ReturnSmoothedTrajNetwork}
\alias{ReturnSmoothedTrajNetwork.SingleCellExperiment}
\alias{ReturnSmoothedTrajNetwork,SingleCellExperiment-method}
\title{Return the milestone network of a trajectory}
\usage{
ReturnSmoothedTrajNetwork.SingleCellExperiment(object, clustering.name)

\S4method{ReturnSmoothedTrajNetwork}{SingleCellExperiment}(object, clustering.name)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{clustering.name}{a character denoting 
the name of the trajectory (clustering).}
}
\value{
a data frame
}
\description{
The function returns the milestone network of a trajectory 
smoothed with Slingshot.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) #Use default N.clusterings=10000
sce <- SelectClusterings(sce,selection.method = 1,selection.N.models = 2)
sce <- RunSmoothing(sce)
clustering_name <-  ReturnTrajNames(sce)[1]
mst_clustering <- ReturnMSTNetwork(sce,clustering_name)
}
\keyword{network}
\keyword{trajectory}
