% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ReturnTrajNames}
\alias{ReturnTrajNames}
\alias{ReturnTrajNames.SingleCellExperiment}
\alias{ReturnTrajNames,SingleCellExperiment-method}
\title{Return trajectory names}
\usage{
ReturnTrajNames.SingleCellExperiment(object)

\S4method{ReturnTrajNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}
}
\value{
a character vector
}
\description{
The function returns the names of the trajectories, 
i.e. those that were selected using the \code{SelectClusterings} function.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce,selection.method = 1,selection.N.models = 2)
ReturnTrajNames(sce)
}
\keyword{clustering}
\keyword{names}
