% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{RunCLARA}
\alias{RunCLARA}
\title{Run k-medoids (CLARA) clustering}
\usage{
RunCLARA(X, ks, min.cluster.size, verbose)
}
\arguments{
\item{X}{a matrix that specifies the embedding (dim.reduced data).}

\item{ks}{a character or numeric denoting the starting cluster.}

\item{min.cluster.size}{a positive integer specifying 
the minimum number of cells the clusters can have. Clusterings that don't
meet this requirement are filtered out. Note that this should be generally
greater or equal to the number of dimensions}

\item{verbose}{a logical for denoting whether the index of the clustering
that is generated should be printed for progress following purposes.}
}
\value{
a list that includes the clustering results
}
\description{
The function runs CLARA algorithm multiple times, with the 
number of clusters (k) corresponding to the parameter \code{ks}, 
a character vector of unique k values (e.g. \code{c("9", "9.1", "9.2")}).
}
\keyword{clara}
\keyword{cluster}
