% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunClustering}
\alias{RunClustering}
\alias{RunClustering.SingleCellExperiment}
\alias{RunClustering,SingleCellExperiment-method}
\title{Run k-medoids clustering}
\usage{
RunClustering.SingleCellExperiment(
  object,
  k.range,
  min.cluster.size,
  N.clusterings,
  verbose
)

\S4method{RunClustering}{SingleCellExperiment}(
  object,
  k.range = 3:20,
  min.cluster.size = 5,
  N.clusterings = 10000,
  verbose = FALSE
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}}

\item{k.range}{a numeric vector specifying 
the range of the number of clusters. E.g. 3:20 would generate clusterings
for which the number of clusters ranges from 3 to 20.}

\item{min.cluster.size}{a positive integer specifying 
the minimum number of cells the clusters can have. Clusterings that don't
meet this requirement are filtered out. Note that this should be generally
greater or equal to the number of dimensions 
in the dimensionally reduced data. Otherwise the Slingshot distance metric
can produce errors.}

\item{N.clusterings}{a positive integer denoting the initial number of 
clustering reuslts. After filtering out those with too small clusters the 
number will go down.}

\item{verbose}{a logical for denoting whether the index of the clustering
that is generated should be printed for progress following purposes.}
}
\value{
an object of class \code{SingleCellExperiment}
}
\description{
The function runs the k-medoids clustering, filters out clusterings that
have too small clusters, and calculates the cell connectivity measure.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
}
\keyword{clara}
\keyword{clustering}
\keyword{k-medoids}
