% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunDimRed}
\alias{RunDimRed}
\alias{RunDimRed.SingleCellExperiment}
\alias{RunDimRed,SingleCellExperiment-method}
\title{Run dimensionality reduction}
\usage{
RunDimRed.SingleCellExperiment(
  object,
  dim.red.method,
  dim.red.features,
  dim.reduction.par.list
)

\S4method{RunDimRed}{SingleCellExperiment}(
  object,
  dim.red.method = "lmds",
  dim.red.features = NULL,
  dim.reduction.par.list = list(ndim = 5)
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}.}

\item{dim.red.method}{a character denoting which dimensionality 
reduction method to use. The options are 
`lmds` (Landmark Multi-Dimensional Scaling), 
`mds` (Multi-Dimensional Scaling),  #' tsne, umap, pca
`tsne` (t-distributed stochastic neighbor embedding), 
`umap` (Uniform Manifold Approximation and Projection), 
`custom` if the user wants to use a custom embedding. The custom
embedding can be provided by 
dim.reduction.par.list = list(dim_red=X), where `X` is the embedding.}

\item{dim.red.features}{a character vector specifying the features
that are used as input in the dimensionality reduction. Useful in situations
where the user wants to use the full gene expression matrix to visualize
gene expression but only some of the features in the dimensionality reduction}

\item{dim.reduction.par.list}{a list specifying the parameters to use
in the dimensionality reduction functions of dyndimred R package. 
For example, list(ndim=5) can be used to specify the number of 
dimensions for lmds and mds.}
}
\value{
an object of class \code{SingleCellExperiment}
}
\description{
The function performs dimensionality reduction using 
\code{dyndimred} package methods.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
}
\keyword{dimensionality}
\keyword{features}
\keyword{reduction}
