% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{RunSlingshotSmoothing}
\alias{RunSlingshotSmoothing}
\title{Run Slingshot smoothing for a clustering}
\usage{
RunSlingshotSmoothing(rd, labels, start.clus, slingshot.par.list)
}
\arguments{
\item{rd}{a matrix that includes the dimensionally reduced data.}

\item{labels}{a vector that includes the base clustering.}

\item{start.clus}{an integer or a character 
for the name of the start cluster.}

\item{slingshot.par.list}{a list that includes the parameters of
Slingshot.}
}
\value{
a list that includes objects of 
classes dynwrap and PseudotimeOrdering
}
\description{
The function runs the Slingshot smoothing for a base clustering.
}
\keyword{slingshot}
\keyword{smoothing}
