% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{RunTotem}
\alias{RunTotem}
\alias{RunTotem.SingleCellExperiment}
\alias{RunTotem,SingleCellExperiment-method}
\title{Run Totem}
\usage{
RunTotem.SingleCellExperiment(
  object,
  dim.red.method,
  dim.red.features,
  dim.reduction.par.list,
  k.range,
  min.cluster.size,
  N.clusterings,
  selection.method,
  selection.N.models,
  selection.k.range,
  selection.stratified,
  prior.clustering,
  slingshot.par.list
)

\S4method{RunTotem}{SingleCellExperiment}(
  object,
  dim.red.method = "lmds",
  dim.red.features = NULL,
  dim.reduction.par.list = list(ndim = 5),
  k.range = 3:20,
  min.cluster.size = 5,
  N.clusterings = 10000,
  selection.method = 1,
  selection.N.models = 1,
  selection.k.range = NULL,
  selection.stratified = FALSE,
  prior.clustering = NULL,
  slingshot.par.list = list(shrink = 1, reweight = TRUE, reassign = TRUE, thresh = 0.001,
    maxit = 10, stretch = 2, smoother = "smooth.spline", shrink.method = "cosine",
    dist.method = "slingshot")
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}.}

\item{dim.red.method}{a character denoting which dimensionality 
reduction method to use. The options are 
`lmds` (Landmark Multi-Dimensional Scaling), 
`mds` (Multi-Dimensional Scaling),  #' tsne, umap, pca, 
`tsne` (t-distributed stochastic neighbor embedding), 
`umap` (Uniform Manifold Approximation and Projection), 
`custom` if the user wants to use a custom embedding. The custom
embedding can be provided by 
dim.reduction.par.list = list(dim_red=X), where `X` is the embedding.}

\item{dim.red.features}{a character vector specifying the features
that are used as input in the dimensionality reduction. Useful in situations
where the user wants to use the full gene expression matrix to visualize
gene expression but only some of the features in the dimensionality reduction}

\item{dim.reduction.par.list}{a list specifying the parameters to use
in the dimensionality reduction functions of dyndimred R package. 
For example, list(ndim=5) can be used to specify the number of 
dimensions for lmds and mds.}

\item{k.range}{a numeric vector specifying 
the range of the number of clusters. E.g. 3:20 would generate clusterings
for which the number of clusters ranges from 3 to 20.}

\item{min.cluster.size}{a positive integer specifying 
the minimum number of cells the clusters can have. Clusterings that don't
meet this requirement are filtered out. Note that this should be generally
greater or equal to the number of dimensions 
in the dimensionally reduced data. Otherwise the Slingshot distance metric
can produce errors.}

\item{N.clusterings}{a positive integer denoting the initial number of 
clusterings that are generated. 
After filtering out those with too small clusters the number will go down.}

\item{selection.method}{a positive integer specifying 
the clustering selection method.
`1` will use the VRC with the cell connectivity measure as input.
`2` will use the VRC with the dimensionally reduced data as input.
`3` will use the mean of `1` and `2`.
`4` will use the mean of `1` and the ARI between a clustering and 
the `prior.clustering`. ARI is the adjusted Rand index that calculates the
similarity of two clusterings. `prior.clustering` parameter specifies the
prior clustering that the user wants to use to direct the clustering search
towards clusterings that are more similar with `prior.clustering`.
`5` will use the mean of `2` and the ARI between a clustering and 
the `prior.clustering`. ARI is the adjusted Rand index that calculates the
similarity of two clusterings. `prior.clustering` parameter specifies the
prior clustering that the user wants to use to direct the clustering search
towards clusterings that are more similar with `prior.clustering`.}

\item{selection.N.models}{a positive integer specifying the number of
top-ranking clusterings that are selected. If the value is a a numeric 
vector, this selects a set of clusterings of which position in the ranked
list matches the integers (e.g., 2:4 would select 
the clusterings ranked 2nd,3rd and 4th).}

\item{selection.k.range}{an integer vector that specifies 
the numbers of clusters that are allowed in the selection. E.g.
c(10,11) would select clusterings that have 10 or 11 clusters.}

\item{selection.stratified}{a logical variable (TRUE or FALSE) that can be
used to use the stratified selection method. If activated, 
the number of clusters will vary evenly in the selected clusterings. Can
be useful if the top results produce trajectories that are not satisfactory,
and the user wants to try very different trajectories.}

\item{prior.clustering}{a vector that includes a prior clustering. When used
in conjunction with `selection.method=4` or `selection.method=5`, it
can be used to select clusterings that are more similar with
the prior clustering.}

\item{slingshot.par.list}{a list of the Slingshot parameters.}
}
\value{
an object of class \code{SingleCellExperiment}
}
\description{
The function wraps all steps in Totem from dimensionality reduction 
to Slingshot smoothing.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)

sce <- SingleCellExperiment(assays = list(
counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- RunTotem(sce,N.clusterings=200) #use default N.clusterings=10000
}
\keyword{Totem}
\keyword{inference}
\keyword{trajectory}
\keyword{wrapper}
