% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SlingshotDistanceMatrix}
\alias{SlingshotDistanceMatrix}
\title{Calculate the Slingshot distance matrix}
\usage{
SlingshotDistanceMatrix(x, centers, clusters)
}
\arguments{
\item{x}{a matrix for the embedding}

\item{centers}{a matrix for the cluster centroids}

\item{clusters}{a vector for the clustering}
}
\value{
a matrix of the cluster distances
}
\description{
The function calculates the Slingshot distance matrix of a clustering.
The method is the covariance-based method (Mahalanobis) used by default 
in Slingshot. The function has been copied from 
the source code of the Slingshot R package.
https://github.com/kstreet13/slingshot
}
\keyword{Mahalanobis}
\keyword{Slingshot}
\keyword{covariance}
\keyword{distance}
\keyword{matrix}
