% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEntities.R
\name{toEntities}
\alias{toEntities}
\title{toEntities}
\usage{
toEntities(data, select = NULL)
}
\arguments{
\item{data}{A data.frame to be converted}

\item{select}{A vector of column names (attributes). Default is to take all columns.}
}
\value{
A data.frame with unique rows of the data table with selected columns with
a new column \code{empirical} denoting the empirical relative frequency distribution.
The attribute \code{attr(*, "N")} corresponds to the number of data points (rows).
}
\description{
Converts a data table to the entity representation
}
