% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VizClustering}
\alias{VizClustering}
\alias{VizClustering.SingleCellExperiment}
\alias{VizClustering,SingleCellExperiment-method}
\title{Visualize a user-provided clustering}
\usage{
VizClustering.SingleCellExperiment(
  object,
  clustering,
  dim.red.type,
  viz.dim.red,
  plot.mst
)

\S4method{VizClustering}{SingleCellExperiment}(
  object,
  clustering = NULL,
  dim.red.type = NULL,
  viz.dim.red = NULL,
  plot.mst = TRUE
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}.}

\item{clustering}{a vector that specifies the clustering.}

\item{dim.red.type}{a character that denotes the name of the 
dimensionality reduction method stored in the \code{SingleCellExperiment}
object.}

\item{viz.dim.red}{a matrix that specifies a custom two-dimensional
embedding for visualization.}

\item{plot.mst}{a logical denoting whether the MST should
be visualized.}
}
\value{
a \code{ggplot2} object
}
\description{
The function can be used to visualize a user-provided clustering.
The MST of the clustering can also be visualized, which is generated
using the Slingshot distance method.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce)
sce <- RunSmoothing(sce)
dim_red <- dyndimred::dimred_mds(t(logcounts(sce)),ndim=2)
clustering_name <- ReturnTrajNames(sce)[1]
clustering <- ReturnClustering(sce,clustering_name)
VizClustering(sce,viz.dim.red=dim_red,clustering)
}
\keyword{clustering}
\keyword{embedding}
\keyword{two-dimensional}
\keyword{visualization}
