% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VizFeatureExpression}
\alias{VizFeatureExpression}
\alias{VizFeatureExpression.SingleCellExperiment}
\alias{VizFeatureExpression,SingleCellExperiment-method}
\title{Visualize feature expression}
\usage{
VizFeatureExpression.SingleCellExperiment(
  object,
  traj.name,
  feature.names,
  dim.red.type,
  viz.dim.red,
  plot.traj
)

\S4method{VizFeatureExpression}{SingleCellExperiment}(
  object,
  traj.name = NULL,
  feature.names = NULL,
  dim.red.type = NULL,
  viz.dim.red = NULL,
  plot.traj = TRUE
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}.}

\item{traj.name}{a character for the trajectory (clustering) name.}

\item{feature.names}{a character vector of the feature names.}

\item{dim.red.type}{a character that specifies the name of the dimensionality
reduction method saved in the \code{SingleCellExperiment} object.}

\item{viz.dim.red}{a matrix for the user-provided, custom dimensionality
reduction.}

\item{plot.traj}{a logical for whether the trajectory should 
be plotted.}
}
\value{
a \code{ggplot2} object
}
\description{
The function visualizes expression of one or multiple features
on an embedding .
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce)
sce <- RunSmoothing(sce)
dim_red <- dyndimred::dimred_mds(t(logcounts(sce)),ndim=2)
clustering_name <- ReturnTrajNames(sce)[1]
feature_names <- rownames(logcounts(sce))[c(1,2)]
VizFeatureExpression(sce,viz.dim.red=dim_red,feature.names=feature_names)
}
\keyword{dimensionality}
\keyword{embedding}
\keyword{expression}
\keyword{feature}
\keyword{reduction}
\keyword{visualization}
