% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VizMST}
\alias{VizMST}
\alias{VizMST.SingleCellExperiment}
\alias{VizMST,SingleCellExperiment-method}
\title{Visualize a Minimum Spanning Tree}
\usage{
VizMST.SingleCellExperiment(
  object,
  clustering.names,
  dim.red.type,
  viz.dim.red
)

\S4method{VizMST}{SingleCellExperiment}(
  object,
  clustering.names = NULL,
  dim.red.type = NULL,
  viz.dim.red = NULL
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}.}

\item{clustering.names}{a character vector that specifies the names of the
MSTs (clusterings) to be visualized.}

\item{dim.red.type}{a character that denotes the name of the 
dimensionality reduction method stored in the \code{SingleCellExperiment}
object.}

\item{viz.dim.red}{a matrix that specifies a custom two-dimensional
embedding for visualization.}
}
\value{
a \code{ggplot2} object
}
\description{
The function visualizes a Minimum Spanning Tree on an embedding. The
MST is generated based on the clusterings of which names are
given through the \code{clustering.names} parameter. The distance method
is the Slingshot distance method, which can not be currently changed.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce)
sce <- RunSmoothing(sce)
dim_red <- dyndimred::dimred_mds(t(logcounts(sce)),ndim=2)
clustering_name <- ReturnTrajNames(sce)[1]
VizMST(sce,viz.dim.red=dim_red,clustering.names=clustering_name)
}
\keyword{MST}
\keyword{visualization}
