% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{VizOneMST}
\alias{VizOneMST}
\title{Visualize an MST}
\usage{
VizOneMST(viz.dim.red, X, clustering, clustering.name, return.traj = FALSE)
}
\arguments{
\item{viz.dim.red}{a matrix of the visualization.}

\item{X}{a matrix that specifies the embedding (dim.reduced data).}

\item{clustering}{a vector of cluster labels.}

\item{clustering.name}{a character, the name of the clustering.}

\item{return.traj}{a logical whether to return the trajectory as well.}
}
\value{
a ggplot2 object
}
\description{
The function visualizes an MST and the underlying clustering.
}
\keyword{MST}
\keyword{visualize}
