% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{VizSmoothedTraj}
\alias{VizSmoothedTraj}
\alias{VizSmoothedTraj.SingleCellExperiment}
\alias{VizSmoothedTraj,SingleCellExperiment-method}
\title{Visualize a smoothed trajectory}
\usage{
VizSmoothedTraj.SingleCellExperiment(
  object,
  traj.names,
  dim.red.type,
  viz.dim.red,
  plot.pseudotime
)

\S4method{VizSmoothedTraj}{SingleCellExperiment}(
  object,
  traj.names = NULL,
  dim.red.type = NULL,
  viz.dim.red = NULL,
  plot.pseudotime = FALSE
)
}
\arguments{
\item{object}{an object of class \code{SingleCellExperiment}.}

\item{traj.names}{a character vector of the trajectory (clustering) names.}

\item{dim.red.type}{a character that specifies the name of the 
dimensionality reduction method saved in the \code{SingleCellExperiment} 
object.}

\item{viz.dim.red}{a matrix for the user-provided, custom dimensionality
reduction.}

\item{plot.pseudotime}{a logical for whether the pseudotime of 
the trajectories should be plotted.}
}
\value{
a \code{ggplot2} object
}
\description{
The function visualizes a Slingshot-smoothed trajectory.
}
\examples{
## DO NOT RUN.  USE THE VIGNETTE CODE TO RUN YOUR ANALYSIS.
library(SingleCellExperiment)
data(binary_tree_1)
sce <- SingleCellExperiment(assays = list(counts=t(binary_tree_1$raw_data),
logcounts = t(binary_tree_1$normalized_data)))
sce <- PrepareTotem(sce)
sce <- RunDimRed(sce,dim.red.method="lmds",
dim.reduction.par.list=list(ndim=5))
sce <- RunClustering(sce,N.clusterings=200) # Use default N.clusterings=10000
sce <- SelectClusterings(sce)
sce <- RunSmoothing(sce)
dim_red <- dyndimred::dimred_mds(t(logcounts(sce)),ndim=2)
clustering_name <- ReturnTrajNames(sce)[1]
VizSmoothedTraj(sce,viz.dim.red=dim_red,traj.names=clustering_name)
}
\keyword{dimensionality}
\keyword{embedding}
\keyword{reduction}
\keyword{smooth}
\keyword{trajectory}
\keyword{visualization}
