% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEntities.R
\name{toEntities}
\alias{toEntities}
\title{toEntities}
\usage{
toEntities(data, select = NULL, cartesian = FALSE)
}
\arguments{
\item{data}{a data.frame to be converted}

\item{select}{a vector of column names (attributes). Default is to take all columns.}

\item{cartesian}{a logical indicating whether the entities should comprise
only the observed tupples of manifestations (\code{FALSE}, default) or
the cartesian product (all possible combinations) of the attribute
manifestations (\code{TRUE})}
}
\value{
A data.frame with unique rows of the data table with selected columns with
a new column \code{empirical} denoting the empirical relative frequency distribution.
The attribute \code{attr(*, "N")} corresponds to the number of data points (rows).
}
\description{
Converts a data table to the entity representation
}
